Move uConsole board init patch to main probe to fix race condition.
The axp20x_battery driver's probe is deferred possibly due to dependence
on the axp22x-adc driver, which itself is deferred. However, the board
initialization was within the axp20x_battery driver, and would apply
too late, leaving the 4G module unable to get power.

The race condition is as follows:
1. The axp20x-regulator driver probes early (~3.5s) and tries to
   enable the 'ldo_io0' rail (for the 4G module), as defined in
   the DTO.
2. The patch that configures the AXP20X_GPIO0_CTRL register to
   set that pin as an LDO doesn't run until the battery driver
   loads (~19.6s).
3. The regulator fails to enable the rail, as the hardware is
   not yet configured.

We fix this by moving the board init patch to the axp20x-i2c driver
where it should load first every time.